feat(calendar): sort properly and replace year headers
